+++ include/GafferImage/OpenColorIOTransform.h
@@ -68,9 +68,18 @@ class GAFFERIMAGE_API OpenColorIOTransform : public ColorProcessor
 
 		GAFFER_GRAPHCOMPONENT_DECLARE_TYPE( GafferImage::OpenColorIOTransform, OpenColorIOTransformTypeId, ColorProcessor );
 
+		/// Returns the OCIO processor for this node, taking into account
+		/// the current Gaffer context and the OCIO context specified by
+		/// `contextPlug()`. Returns nullptr if this node is a no-op.
+		OpenColorIO::ConstProcessorRcPtr processor() const;
+		/// Returns a hash that uniquely represents the result of calling
+		/// `processor()` in the current context.
+		IECore::MurmurHash processorHash() const;
+
 		/// Derived classes must implement this to return a valid OpenColorIO
 		/// Transform which can be used by an OpenColorIO Processor or a null
 		/// pointer if no processing should take place.
+		/// \todo Make protected again.
 		virtual OpenColorIO::ConstTransformRcPtr transform() const = 0;
 
 	protected :