+++ python/GafferUI/ErrorDialogue.py
@@ -75,9 +75,8 @@ class ErrorDialogue( GafferUI.Dialogue ) :
 				)
 
 			if messages is not None :
-				messageWidget = GafferUI.MessageWidget()
-				for m in messages :
-					messageWidget.messageHandler().handle( m.level, m.context, m.message )
+				messageWidget = GafferUI.MessageWidget( toolbars = True )
+				messageWidget.setMessages( messages )
 
 			if details is not None :
 				with GafferUI.Collapsible( label = "Details", collapsed = True ) :