+++ python/GafferUI/PlugValueWidget.py
@@ -425,13 +425,14 @@ class PlugValueWidget( GafferUI.Widget ) :
 				}
 			)
 
-		if any( Gaffer.NodeAlgo.presets( p ) for p in self.getPlugs() ) :
-			menuDefinition.append(
-				"/Preset", {
-					"subMenu" : Gaffer.WeakMethod( self.__presetsSubMenu ),
-					"active" : self._editable()
-				}
-			)
+		with self.getContext() :
+			if any( Gaffer.NodeAlgo.presets( p ) for p in self.getPlugs() ) :
+				menuDefinition.append(
+					"/Preset", {
+						"subMenu" : Gaffer.WeakMethod( self.__presetsSubMenu ),
+						"active" : self._editable()
+					}
+				)
 
 		if len( menuDefinition.items() ) :
 			menuDefinition.append( "/LockDivider", { "divider" : True } )
@@ -639,7 +640,6 @@ class PlugValueWidget( GafferUI.Widget ) :
 		# any which aren't available for all plugs.
 		result = IECore.MenuDefinition()
 		for presetName in presets :
-
 			menuPath = presetName if presetName.startswith( "/" ) else "/" + presetName
 			result.append(
 				menuPath, {
@@ -653,9 +653,10 @@ class PlugValueWidget( GafferUI.Widget ) :
 
 	def __applyPreset( self, presetName, *unused ) :
 
-		with Gaffer.UndoScope( next( iter( self.getPlugs() ) ).ancestor( Gaffer.ScriptNode ) ) :
-			for p in self.getPlugs() :
-				Gaffer.NodeAlgo.applyPreset( p, presetName )
+		with self.getContext() :
+			with Gaffer.UndoScope( next( iter( self.getPlugs() ) ).ancestor( Gaffer.ScriptNode ) ) :
+				for p in self.getPlugs() :
+					Gaffer.NodeAlgo.applyPreset( p, presetName )
 
 	def __applyReadOnly( self, readOnly ) :